ft_malloc

+++ libft_malloc/bonus/srcs/block_related/get_avail_block_bonus.c
@@ -8,6 +8,8 @@ static void	*first_fit(t_heap_group *heap, size_t size)
 	free_block = heap->free_block;
 	while (free_block && free_block->size < size)
 		free_block = free_block->next;
+	if (!free_block)
+		return NULL;
 	heap->free_block = delete_block(free_block, size, 1);
 	block = add_alloc_block(heap, free_block);
 	return (void *)((char *)block + sizeof(t_block));
